fix: move CopyResult type to shared types.ts (copier.ts was deleted)
