+++ extensions/magic-ingest/src/utils/types.ts
@@ -1,0 +1,6 @@
+export interface CopyResult {
+  sourcePath: string;
+  destPath: string;
+  success: boolean;
+  error?: string;
+}
